requirements.txt
# For linux use pip

# For linux use conda, but split instalation into 2 commands shown bellow.
# Every time you want to use a one of those commands comment libraries under another.

# conda install --file requirements.txt -c pytorch -c nvidia -c anaconda -c conda-forge
python==3.10
opencv==4.6.0
pandas==1.5.2
pytorch==1.12.1
torchvision==0.13.1
torchaudio==0.12.1
torchmetrics==0.9.3
patool==1.12
py-xgboost==1.7.3
pytorch-lightning==1.7.3
wandb==0.14.0
cudatoolkit==11.6
pillow==9.0.0
gdown==4.6.4
# conda install --file requirements.txt -c defaults -f
# or just
# conda install intel-openmp -c defaults -f
intel-openmp==2021.4.0